+++ CHANGELOG.md
@@ -11,9 +11,12 @@
 ### 🛠  Improvements
 - [624](https://github.com/vegaprotocol/data-node/issues/624) - Support subscriptions in new `API`
 - [666](https://github.com/vegaprotocol/data-node/issues/666) - Cache latest market data
+- [564](https://github.com/vegaprotocol/data-node/issues/564) - Add cursor based pagination to market data requests
+- [](https://github.com/vegaprotocol/data-node/issues/xxx) -
 
 ### 🐛 Fixes
 - [657](https://github.com/vegaprotocol/data-node/issues/657) - Add missing creation field in `ERC20` withdrawal bundle
+- [668](https://github.com/vegaprotocol/data-node/issues/668) - Ensure entity wrappers always hold timestamps to microsecond resolution 
 - [662](https://github.com/vegaprotocol/data-node/issues/662) - Fix auction trigger enum lookup
 
 ## 0.51.1